requirements.txt
SQLAlchemy~=1.4.36
fastapi~=0.78.0
PyJWT~=2.4.0
pydantic~=1.9.1
uvicorn~=0.17.6